1. Adjust relative parameter. 2. Change font size from int to float.

+++ osu.Game.Rulesets.Karaoke/Skinning/Components/FontInfo.cs
@@ -9,7 +9,7 @@ namespace osu.Game.Rulesets.Karaoke.Skinning.Components
 
         public bool Bold { get; set; }
 
-        public int CharSize { get; set; }
+        public float CharSize { get; set; }
 
         public int EdgeSize { get; set; }
     }
